% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctt.R
\name{alpha.scan}
\alias{alpha.scan}
\title{Alpha Cronback scan}
\usage{
alpha.scan(
  data,
  items,
  gain.format = TRUE,
  digits = 3,
  parallel = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{data}{a data.frame containing the test or scale response in columns}

\item{items}{The name of the items (should be >1 )}

\item{item.drop}{}
}
\description{
Alpha Cronback scan
}
