% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apa.R
\name{apa.n}
\alias{apa.n}
\title{APA formating number}
\usage{
apa.n(x, d = 2, oto = F, p = F)
}
\arguments{
\item{x}{a numric}

\item{d}{the number of digits needed}

\item{oto}{a true of false statement (false by default) to indicate if the value x  of the nurmric value provided would matematicly constrict between -one to +one a numric}

\item{p}{a true of false  statement (false by default) to indicate if the value x is a p-value}
}
\value{
a formated number for apa normed document
}
\description{
convert a numerical result to an apa norme restul
}
\examples{
apa.n(34.036, d=2) ;
apa.n(34.036, d=3)
apa.n(34.036, d=4)
apa.n(0.001, d=1)
apa.n(0.001, d=2)
apa.n(0.001, d=3)
apa.n(0.001, d=2,p=T)
apa.n(0.001, d=3,p=T)
apa.n(-0.10, d=2,oto=T)
apa.n(-0.10, d=3,oto=T)
}
