% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{crosslag_model}
\alias{crosslag_model}
\title{Crosslag model formulat}
\usage{
crosslag_model(
  variables,
  times,
  order = "times/variables",
  join = "_",
  prefix = NULL
)
}
\arguments{
\item{variables}{A vector of names variables}

\item{times}{names of the time indicator for the model}

\item{order}{a string defining the order of the variable and the time indicator by default "times/variables" or "variables/times"}

\item{join}{a string defining by what carractere are join variable and time (by default "_")}
}
\value{
return a lavaan formulat for a crosslag model
}
\description{
Create a Lavaan modem formulat for a classic crosslag model
}
\examples{
vars = c("x","y")
t = 1:4
crosslag_model(variables = vars, times = t, order = "times/variables", join="_" )
crosslag_model(variables = vars, times = t, order = "variables/times", join="" )

}
