% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFA_comp.R
\name{EFA.comp}
\alias{EFA.comp}
\title{EFA comparaison model}
\usage{
EFA.comp(data, items, nfactor, parallel = TRUE, fit = TRUE, ...)
}
\arguments{
\item{data}{e}

\item{items}{e}

\item{nfactor}{e}

\item{parallel}{e}

\item{fit}{e}
}
\value{
test
}
\description{
Function that provide model fit for difference EFA model
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
eval1 = EFA.comp(data = inference, items = items, nfactor = 1:3)
summary(eval1)


}
