% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFA_explor.R
\name{EFA.explo}
\alias{EFA.explo}
\title{FA.exploratory}
\usage{
EFA.explo(data, items, plot = FALSE, ...)
}
\arguments{
\item{data}{a data.frame containing the test or scale response in columns}

\item{items}{The name of the items}

\item{plot}{Plot vss and parallel analysis}

\item{...}{}
}
\value{
the result provide a print of the factor detected for each of the method
}
\description{
The function provide information factor analysis and if the actual data are recommended for factoriale analysis. 
It also used multiple statisical function to provide an indication regarding the number of factor present in the dataset.
Those criteria containt the parallele analysis method, the Velicer's MAP, VSS and Kaiser criterion.
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
eval1 <- EFA.explo(inference, items)
eval1
}
\seealso{
\code{\link{fa.parallel}}  \code{\link{vss}}
}
