% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.ctt.R
\name{eval.ctt}
\alias{eval.ctt}
\title{Classical Test Theory Summary}
\usage{
eval.ctt(
  data,
  items,
  data.response = NULL,
  items.rar = NULL,
  keys = NULL,
  digits = 3,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{data}{a data.frame containing the test or scale response in columns}

\item{items}{The name of the items (should be >1 )}

\item{digits}{number of digits for result}

\item{...}{}
}
\value{
a summary of the result and a list with the complet result of each function
}
\description{
the fonction provide a warpping of the \code{\link{scan.freq}}, \code{\link{scan.diff}}, \code{\link{scan.disc}}\code{\link{scan.alpha}} and \code{\link{scan.rar}} functions
}
\examples{
library(psychometRics)
library(dplyr)

data("inference")
items <- c(paste0("item_0",3:9),"item_10")
ev1 <- eval.ctt(data = inference, items = items)
ev1


item.1 <- c(rep("A", 15), rep("B", 10), rep("C", 15))
item.2 <- c(rep("A", 10), rep("B", 15), rep("C", 15))
item.3 <- c(rep("A", 5), rep("B", 15), rep("C", 20))
item.4 <- c(rep("A", 3), rep("B", 19), rep("C", 18))
item.5 <- c(rep("A", 11), rep("B", 13), rep("C", 16))

dat.response <- data.frame(
  item.1 = item.1,
  item.2 = item.2,
  item.3 = item.3,
  item.4 = item.4,
  item.5 = item.5
)

key <- c("A","B","C","A","B")

dat.score <- data.frame()
for (i in 1:nrow(dat.response)){
  tmp <- dat.response[i,] == key
  dat.score <- rbind(dat.score, as.data.frame(tmp))
}

dat.score <- as.data.frame(sapply(dat.score,as.numeric))
ev2 <- eval.ctt(dat.score,items = paste0("item.",1:5),data.response = dat.response,items.rar = paste0("item.",1:5),keys = key)
ev2
}
\seealso{
\code{\link{scan.freq}}, \code{\link{scan.diff}}, \code{\link{scan.disc}}\code{\link{scan.alpha}} and \code{\link{scan.rar}}
}
