% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{fa.CFI}
\alias{fa.CFI}
\title{CFI indicator}
\usage{
fa.CFI(x)
}
\arguments{
\item{x}{a  psych:fa class}
}
\value{
the CFI index of the tested model
}
\description{
extract a CFI fit index from a psych:fa class result
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
FA <- psych::fa(inference[,items], nfactors = factor)
fa.CFI(FA)
}
