% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{factor_comparison2}
\alias{factor_comparison2}
\title{Factor comparison}
\usage{
factor_comparison2(data, items, nfactors, ...)
}
\arguments{
\item{data}{a data.frame containing the test or scale response in columns}

\item{items}{The name of the items}

\item{factors}{a facto class object from factor_explorer() or a vector with the number of factor}
}
\value{
a table with a CHI² comparison and CFI, TLI, RMSEA and SRMR differences
}
\description{
Compute anova and fit index to compare multiple factor analyse fit for multiple factors solution
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
f <- factor_explorer(inference, items)
factor_comparison2(inference, items, f)
factor_comparison2(inference, items, c(1,2,3))
}
