% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZZfactorial.R
\name{factor_explorer}
\alias{factor_explorer}
\title{Factor explorer}
\usage{
factor_explorer(data, items, plot = FALSE, ...)
}
\arguments{
\item{data}{a data.frame containing the test or scale response in columns}

\item{items}{The name of the items}

\item{plot}{Plot vss and parallel}

\item{...}{}
}
\value{
the result provide a print of the factor detected for each of the method
}
\description{
the function use parallel analyse, Velicer and VSS method to provide the possible number of factor.
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
factor_explorer(inference, items)
}
