% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_alpha.R
\name{print.alpha}
\alias{print.alpha}
\title{Print method for the alpha class}
\usage{
\method{print}{alpha}(x)
}
\arguments{
\item{x}{An alpha class object}
}
\value{
print the result of the function
}
\description{
Print method for the alpha class
}
