% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_diff.R
\name{print.difficulty}
\alias{print.difficulty}
\title{Print method for the difficulty class}
\usage{
\method{print}{difficulty}(x)
}
\arguments{
\item{x}{An alpha class object}
}
\value{
print the result of the function
}
\description{
Print method for the difficulty class
}
