% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_disc.R
\name{print.discrimination}
\alias{print.discrimination}
\title{Print method for the difficulty class}
\usage{
\method{print}{discrimination}(x)
}
\arguments{
\item{x}{An alpha class object}
}
\value{
print the result of the function
}
\description{
Print method for the difficulty class
}
