% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFA_comp.R
\name{print.EFA.comp}
\alias{print.EFA.comp}
\title{Print method for the rar class}
\usage{
\method{print}{EFA.comp}(x)
}
\arguments{
\item{x}{An alpha class object}
}
\value{
print the result of the function
}
\description{
Print method for the rar class
}
