% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFA_explor.R
\name{print.EFA.explo}
\alias{print.EFA.explo}
\title{Facto print method}
\usage{
\method{print}{EFA.explo}(x)
}
\arguments{
\item{x}{a facto class object}
}
\description{
formating result for the facto class
}
