% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{print.factor.comparaison}
\alias{print.factor.comparaison}
\title{Print method for factor_comparaison}
\usage{
\method{print}{factor.comparaison}(x)
}
\arguments{
\item{x}{a factor.comparaison class}
}
\description{
Print method for factor_comparaison
}
