% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_freq.R
\name{print.frequency}
\alias{print.frequency}
\title{Print method for the frequency class}
\usage{
\method{print}{frequency}(x)
}
\arguments{
\item{x}{An alpha class object}
}
\value{
print the result of the function
}
\description{
Print method for the frequency class
}
