% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_RAR.R
\name{print.rar}
\alias{print.rar}
\title{Print method for the rar class}
\usage{
\method{print}{rar}(x)
}
\arguments{
\item{x}{An alpha class object}
}
\value{
print the result of the function
}
\description{
Print method for the rar class
}
