% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_RI.R
\name{RI_crosslag_model}
\alias{RI_crosslag_model}
\title{The random intercept crosslag panel model}
\usage{
RI_crosslag_model(variables, times, order = "times/variables", join = "_")
}
\arguments{
\item{join}{}
}
\description{
The random intercept crosslag panel model
}
