% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_alpha.R
\name{scan.alpha}
\alias{scan.alpha}
\title{Cronbach alpha scan}
\usage{
scan.alpha(data, items, digits = 3, parallel = FALSE, verbose = TRUE, ...)
}
\arguments{
\item{data}{A data.frame containing participant scores for each item}

\item{items}{The names of the items (should be >1)}

\item{digits}{Number of digits in the table summary}

\item{parallel}{Bool: use parallel backend (doSNOW) when there are many items}

\item{verbose}{Bool for progress visualization}

\item{...}{arguments passed to \code{\link{cronbach.alpha}} from the ltm package}
}
\value{
A data.frame with alpha variations and a z-test per item
}
\description{
The function provides the variation of reliability (Cronbach's alpha)
when an item is dropped. It also computes a z-test comparing alpha with and without
the item. The result is a table with the alpha when the item is removed, the
alpha gain (difference between the alpha with all items and without the item),
the z value and the p value. Positive alpha gain indicates alpha decreases when
the item is removed; negative alpha gain indicates alpha increases when the
item is removed. In general, alpha gain should be positive; if negative, the
item may be unreliable and considered for removal. The function supports
parallel execution when appropriate.
}
\details{
This function provides the variation of Cronbach's alpha when an item
is dropped from the data.
}
\examples{
data("inference")
items <- c(paste0("item_0", 3:9), "item_10")
alp1 <- scan.alpha(data = inference, items = items)
alp1
alp2 <- scan.alpha(inference, items, parallel = TRUE, verbose = FALSE)
alp2

}
\seealso{
\code{\link{eval.ctt}}, \code{\link{cronbach.alpha}}
}
