% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_diff.R
\name{scan.diff}
\alias{scan.diff}
\title{Difficulty scan}
\usage{
scan.diff(data, items, verbose = T)
}
\arguments{
\item{data}{A data.frame containing participant score for each item}

\item{items}{The names of the items (should be >1)}
}
\value{
data.frame with item names and their difficulty
}
\description{
Difficulty scan provides an indicator of item difficulty in classical test theory.
The difficulty is calculated as the item mean score divided by
the difference between the maximum and minimum score of the item.
}
\examples{
data("inference")
items <- names(inference)
dif.scan(data = inference, items = items)

inference[1,1] <- 4
scan.diff(inference, items)
}
\seealso{
\code{\link{eval.ctt}}
}
