% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_disc.R
\name{scan.disc}
\alias{scan.disc}
\title{Discrimination correlation}
\usage{
scan.disc(
  data,
  items,
  type = "rir",
  method = c("cor", "somer", "polyserial"),
  conf.level = 0.95,
  args.cor = list(),
  args.somer = list(),
  args.polyserial = list()
)
}
\arguments{
\item{data}{A data.frame containing participant score for each item}

\item{items}{The names of the items (should be >1)}

\item{type}{Type of correlation: 'rir' for rest-item correlation or 'rit' for total-item correlation (can be both)}

\item{method}{The type(s) of correlation to use: c("cor","somer","polyserial")}

\item{conf.level}{The confidence level for the correlation}

\item{...}{Argument for the cor.test() function}
}
\value{
a data.frame with the result and confidence interval
}
\description{
The function provides item-rest (RIR) or item-total (RIT) correlations for all items in the test.
Multiple types of item correlation are provided such as Somers' D, point-biserial correlation, and polyserial correlation for non-dichotomous items.
Results can be interpreted as the item's ability to discriminate between lower- and higher-performing respondents with respect to the overall raw score.
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
scan.disc(data = inference, item = items, type ="rir")
scan.disc(data = inference, item = items, type ="rit")

}
\seealso{
\code{\link{eval.ctt}}
}
