% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scan_freq.R
\name{scan.freq}
\alias{scan.freq}
\title{Frequency scan}
\usage{
scan.freq(data, items, verbose = T)
}
\arguments{
\item{data}{A data.frame containing participant score for each item}

\item{items}{The names of the items (should be >1)}
}
\value{
data.frame with item names and their frequency of correct responses
}
\description{
Frequency scan provides the frequency of correct answers for dichotomous items (0 or 1) in a data frame
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
freq.scan(inference, items)

inference[1,1] <- 4
scan.freq(inference, items)
}
\seealso{
\code{\link{eval.ctt}}
}
