% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{SEM_cov}
\alias{SEM_cov}
\title{Title}
\usage{
SEM_cov(variables, variance = T)
}
\arguments{
\item{variables}{A vector of names variables}

\item{variance}{If TRUE variance of variable will be writeen}
}
\value{
return a string of model format for lavaan
}
\description{
Title
}
\examples{
SEM_cov(c("A","B","C","D"))
SEM_cov(c("A","B","C","D"), variance = FALSE)

}
