% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM.R
\name{SEM_reg}
\alias{SEM_reg}
\title{SEM regression model}
\usage{
SEM_reg(y, x, latent = FALSE)
}
\arguments{
\item{y}{A vector of string who contain the name of the predicted variable}

\item{x}{A vector of string who contain the name of the predictor variable}
}
\value{
A string formula for a Lavaan model
}
\description{
SEM regression model
}
\examples{

SEM_reg("Y","x")
SEM_reg("Y",c("x1","x2","x3"))
SEM_reg(c("Y1","Y2","Y3"),c("x1","x2","x3"))
SEM_reg("Y",c("x1","x2","x3"), latent=T)


}
