% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SEM_RI.R
\name{SEM_reg_RI}
\alias{SEM_reg_RI}
\title{Regression Random Intercept}
\usage{
SEM_reg_RI(y, x)
}
\arguments{
\item{y}{a variable name for the random intercept}

\item{x}{the variables to be constrained to the random intercept}
}
\description{
Regression Random Intercept
}
