% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval.ctt.R
\name{summary.eval.ctt}
\alias{summary.eval.ctt}
\title{Print Method for CTT_summary}
\usage{
summary.eval.ctt(x)
}
\arguments{
\item{x}{}
}
\value{
print the result of the function
}
\description{
Print Method for CTT_summary
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
t <- CTT_summary(inference, items)
print(t)

}
