% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted.mean2.R
\name{weighted.mean2}
\alias{weighted.mean2}
\title{weighted mean and weighted standard deviation}
\usage{
weighted.mean2(group.mean, pourcent_pop, pourcent_gp)
}
\arguments{
\item{vector}{percentage of each social group in the sample}
}
\value{
the weighted mean and weighted standard deviation
}
\description{
weighted mean and weighted standard deviation
}
\examples{
1+1
group.mean <- c(10,23,15,16)
pop_perc <- c(0.25,0.19,0.5,0.45)
group_perc <- c(0.10,0.25,0.12)
weighted.mean2(group.mean,pop_perc,group_perc)

}
