% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctt.R
\name{CTT_summary}
\alias{CTT_summary}
\title{Classical Test Theory Summary}
\usage{
CTT_summary(data, items, digits = 3, ...)
}
\arguments{
\item{data}{a data.frame containing the test or scale response in columns}

\item{items}{The name of the items (should be >1 )}

\item{...}{}
}
\description{
Classical Test Theory Summary
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
CTT_summary(inference, items)
}
