% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctt.R
\name{discrimination_cor}
\alias{discrimination_cor}
\title{Discrimination correlation}
\usage{
discrimination_cor(
  data,
  items,
  type = "rir",
  method = c("cor", "somer", "polyserial"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{data}{a data.frame containing the test or scale response in columns}

\item{items}{The name of the items (should be >1 )}

\item{type}{type correlation could be rir for Rest Item-correlation or rit for Total Item-correlation (can be both)}

\item{method}{the type of correlation to use c("cor","somer","polyserial")}

\item{conf.level}{the confidence interval for the correlation needed}
}
\value{
a data.frame with the result and confidence interval
}
\description{
The function provide correlation item rest (rir) or correlation item test (rit) for a all of the item in the test.
Multiple type of item corrleation are provided as somer's d, point biserial correlation and polyserial correlation for non dichotomic item
Result could be interpretend as the item ability to discriminate between a lower and higher difficulty for this item in regard to the overall raw score.
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
discrimination_cor(inference, items)
}
