% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctt.R
\name{discrimination_cor}
\alias{discrimination_cor}
\title{Discrimination correlation}
\usage{
discrimination_cor(
  data,
  items,
  type = "rir",
  method = c("cor", "somer", "polyserial"),
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{data}{a data.frame containing the test or scale response in columns}

\item{items}{The name of the items (should be >1 )}

\item{type}{type correlation could be rir for Rest Item-correlation or rit for Total Item-correlation}
}
\value{
a data.frame with the result and confidence interval of pearson correlation and somer delta
}
\description{
Discrimination correlation
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
discrimination_cor(inference, items)
}
