% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctt.R
\name{freq_scan}
\alias{freq_scan}
\title{Frequency scam}
\usage{
freq_scan(data, items)
}
\arguments{
\item{data}{a data.frame containing the test or scale response in columns}

\item{items}{The name of the items (should be >1 )}
}
\value{
data table with the item name and their frequency of correct response
}
\description{
Frequency scan provide the frequency of correct answer for dichotomic item in a data frame
}
\examples{

1+1
}
