% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factorial.R
\name{print.factor.comparaison}
\alias{print.factor.comparaison}
\title{Print method for factor_comparaison}
\usage{
\method{print}{factor.comparaison}(x)
}
\arguments{
\item{x}{a factor.comparaison class}
}
\value{
return a print for the factor_comparaison() result
}
\description{
Print method for factor_comparaison
}
\examples{
data("inference")
items <- c(paste0("item_0",3:9),"item_10")
f <- factor_explorer(inference, items)
ff <- factor_comparison(inference, items, f)
print(ff)
}
